test/system: Fix ulimit tests on Ubuntu host
ulimit is a shell builtin, so the shell it is run from changes
the version that is used. Ubuntu (and Debian) use dash as
their system shell, instead of bash. The means that toolbox
run also used dash with Ubuntu containers, caused an
unexpected version of ulimit to be used in tests. Tests were
updated to force the use of bash and its version of ulimit.

Signed-off-by: Penn Bauman <me@pennbauman.com>